Convert Gradle script plugin for generating UniFfi bindings into a composite build.
