--- buildSrc/build.gradle.kts
@@ -1,8 +1,0 @@
-plugins {
-    `kotlin-dsl`
-    // id("org.gradle.kotlin.kotlin-dsl") version "2.2.0"
-}
-
-repositories {
-    mavenCentral()
-}
