+++ plugins/settings.gradle.kts
@@ -1,0 +1,8 @@
+dependencyResolutionManagement {
+    repositories {
+        mavenCentral()
+        google()
+    }
+}
+
+include(":plugins")
