+++ plugins/src/main/kotlin/org/bitcoindevkit/plugins/Enums.kt
