+++ settings.gradle.kts
@@ -1,3 +1,4 @@
 rootProject.name = "bdk-kotlin"
 
 include(":jvm", ":android")
+includeBuild("plugins")
